+++ ebustoolbox/models.py
@@ -1,15 +1,16 @@
 import shutil
 from datetime import timedelta, datetime
-from pathlib import Path
-import numpy as np
 from functools import partial
+from pathlib import Path
 
+import numpy as np
 from django.conf import settings
 from django.contrib.auth.models import User
 from django.contrib.gis.db import models
 from django.contrib.postgres.fields import ArrayField
-from django.core.serializers.json import DjangoJSONEncoder
+from django.db.models import QuerySet, Sum
 from django.dispatch import receiver
+from django.utils.timezone import make_aware
 
 MINIMAL_TRIP_DURATION_S = 60  # seconds
 
@@ -327,15 +328,6 @@ class Vehicle(models.Model):
         return self.name
 
 
-# ToDo Deprecated
-class VehicleProperties(models.Model):
-    scenario = models.ForeignKey(Scenario, null=True, on_delete=models.CASCADE)
-
-    date = models.DateTimeField()
-    vehicle = models.ForeignKey(Vehicle, on_delete=models.CASCADE)
-    soc = models.FloatField(null=True)
-
-
 class Rotation(models.Model):
     """
     Model representing a rotation associated with a scenario.
@@ -379,6 +371,21 @@ class Rotation(models.Model):
     # SimBA specific data to make SimBA simulations reproducible
     vehicle = models.ForeignKey(Vehicle, on_delete=models.SET_DEFAULT, default=None, null=True)
     allow_opportunity_charging = models.BooleanField(default=None, null=False)
+
+    def get_distance(self):
+        return Route.objects.filter(trip__rotation=self).aggregate(Sum("distance"))
+
+
+def annotate_distance(query: QuerySet[Rotation]):
+    return query.annotate(distance=Sum("trip__route__distance"))
+
+
+def get_longest_distance_rotation(filter_dict: dict) -> Rotation:
+    return annotate_distance(Rotation.objects.filter(**filter_dict)).order_by("distance").last()
+
+
+def get_shortest_distance_rotation(filter_dict: dict) -> Rotation:
+    return annotate_distance(Rotation.objects.filter(**filter_dict)).order_by("distance").first()
 
 
 class EnumVoltageLevel(models.TextChoices):
@@ -458,7 +465,7 @@ class Temperatures(models.Model):
         ...     data=[temperature1, temperature2],
         ... )
         >>> temperatures_instance.save()
-        >>> interpolated_temperature = temperatures_instance.get_temperature(datetime3)
+        >>> interpolated_temperature = temperatures_instance.get_interpolated_temperature(datetime3)
     """
 
     class Meta:
@@ -470,21 +477,24 @@ class Temperatures(models.Model):
 
     # Should the datetime be interpreted as datetime or only time.
     use_only_time = models.BooleanField(null=False, default=True)
-    # two list with datetimes and data. Storing as dict does not seem easily possible, since
-    # serialization of keys as datetime fails but not as values
-    datetimes = models.JSONField(default=list, null=False, encoder=DjangoJSONEncoder)
-    data = models.JSONField(default=list, null=False)
+    # datetimes and associated data
+    datetimes = ArrayField(models.DateTimeField(), default=list)
+    data = ArrayField(models.FloatField(), default=list)
+
     temperature_interpolation = None
+    temperature_closest_function = None
+
+    def make_aware(self):
+        aware_datetimes = []
+        for date_time in self.datetimes:
+            aware_datetimes.append(make_aware(date_time))
+        self.datetimes = aware_datetimes
 
     def save(self, *args, **kwargs):
         """
         Overrides the save method to perform data validation and initialize the
         interpolation function.
         """
-        # Check proper formatting of the data field
-        assert_is_type(self.data, list)
-        if not all([isinstance(obj, datetime) for obj in self.datetimes]):
-            raise AttributeError(f"Data of {self} contains keys, which are not of type Datetime")
         if self.use_only_time:
             dates = {(date.year, date.month, date.day) for date in self.datetimes}
             if len(dates) != 1:
@@ -495,9 +505,10 @@ class Temperatures(models.Model):
         self.temperature_interpolation = get_datetime_interpolation_function(
             self.datetimes, self.data
         )
+        self.temperature_closest_function = get_datetime_closest_function(self.datetimes, self.data)
         super().save(*args, **kwargs)
 
-    def get_temperature(self, datetime: datetime):
+    def get_interpolated_temperature(self, date_time: datetime):
         """
         Retrieves the interpolated temperature for a given datetime.
         """
@@ -506,12 +517,64 @@ class Temperatures(models.Model):
                 self.datetimes, self.data
             )
         if self.use_only_time:
-            date = next(iter(self.datetimes))
-            datetime = datetime.replace(year=date.year, month=date.month, day=date.day)
-        return self.temperature_interpolation(datetime)
+            date = date_time.fromisoformat(str(next(iter(self.datetimes))))
+            date_time = date_time.replace(year=date.year, month=date.month, day=date.day)
+        return self.temperature_interpolation(date_time)
+
+    def get_closest_temperature(self, date_time: datetime):
+        """
+        Retrieves the interpolated temperature for a given datetime.
+        """
+        if self.temperature_closest_function is None:
+            self.temperature_closest_function = get_datetime_closest_function(
+                self.datetimes, self.data
+            )
+        if self.use_only_time:
+            date = date_time.fromisoformat(str(next(iter(self.datetimes))))
+            date_time = date_time.replace(year=date.year, month=date.month, day=date.day)
+        return self.temperature_closest_function(date_time)
 
 
-def get_datetime_interpolation_function(datetimes: list, data: list):
+def get_datetime_closest_function(datetimes: list[datetime], data: list):
+    """
+    Generates a function which returns the closest value based on input datetimes and corresponding data.
+
+    Args:
+        datetimes (list): A list of datetime objects.
+        data (list): A list of values corresponding to the datetimes.
+
+    Returns:
+        function: A datetime interpolation function.
+    """
+    # sort by key
+    sort_index = np.argsort(np.array(datetimes), axis=0)
+    sorted_data = (np.array([datetimes, data]).T)[sort_index]
+    first_time = sorted_data[0, 0]
+    # create the timedelta
+    xp = sorted_data[:, 0] - first_time
+    # cast the timedelta to seconds as int64
+    xp = xp.astype("timedelta64[s]").view("int64")
+    # cast the data to floats
+    fp = sorted_data[:, 1].astype(float)
+    # create the interpolation function
+    partial_closest = partial(np.searchsorted, a=xp, side="left")
+
+    def partial_closest_function(date_time: datetime) -> float:
+        delta_time_as_int = (
+            (np.array([date_time]) - first_time).astype("timedelta64[s]").view("int64")
+        )
+        # idx 1 is the index of xp, where xp[i-1] < delta_time_as_int <= xp[i]
+        idx1 = partial_closest(v=delta_time_as_int)
+        x1 = xp[idx1]
+        idx0 = np.max((idx1 - 1, np.zeros(len(idx1))), axis=0).astype(int)
+        x0 = xp[idx0]
+        indicies = idx1 - (x1 - delta_time_as_int >= delta_time_as_int - x0)
+        return fp[np.max((indicies, np.zeros(len(idx1))), axis=0).astype(int)].squeeze()
+
+    return partial_closest_function
+
+
+def get_datetime_interpolation_function(datetimes: list[datetime], data: list):
     """
     Generates a datetime interpolation function based on input datetimes and corresponding data.
 
@@ -522,9 +585,6 @@ def get_datetime_interpolation_function(datetimes: list, data: list):
     Returns:
         function: A datetime interpolation function.
     """
-
-    # cast datetimes, which are possibly strings to datetimes
-    datetimes = [datetime.fromisoformat(str(x)) for x in datetimes]
     # sort by key
     sort_index = np.argsort(np.array(datetimes), axis=0)
     sorted_data = (np.array([datetimes, data]).T)[sort_index]
@@ -542,8 +602,7 @@ def get_datetime_interpolation_function(datetimes: list, data: list):
         delta_time_as_int = (
             (np.array([date_time]) - first_time).astype("timedelta64[s]").view("int64")
         )
-        if len(delta_time_as_int) == 1:
-            return partial_interp(*delta_time_as_int)
+        delta_time_as_int = delta_time_as_int.squeeze()
         return partial_interp(delta_time_as_int)
 
     return partial_interpolation_function
@@ -607,7 +666,7 @@ class Station(models.Model):
         db_table = "Station"
 
     # Map Engine models need geom and name as first columns
-    geom = models.PointField(dim=3, srid=4326)  # without z elevation
+    geom = models.PointField(dim=3, srid=4326, null=True)  # without z elevation
     name = models.TextField(null=False)
     name_short = models.TextField(null=True, blank=True)
     scenario = models.ForeignKey(Scenario, null=False, on_delete=models.CASCADE)
@@ -808,7 +867,7 @@ class Trip(models.Model):
         departure_time (DateTimeField): The departure time for the trip. Cannot be blank.
         arrival_time (DateTimeField): The arrival time for the trip. Cannot be blank.
 
-        type (str): The type of the trip. Choices defined by EnumTripType.
+        trip_type (str): The type of the trip. Choices defined by EnumTripType.
                     Defaults to EnumTripType.PASSENGER_TRIP.
 
         loaded_mass (float, optional): The mass that is loaded in [kg], i.e. mass of passengers.
@@ -835,7 +894,7 @@ class Trip(models.Model):
         ...     rotation=rotation_instance,
         ...     departure_time=datetime.datetime(2024, 1, 1, 8, 0, 0),
         ...     arrival_time=datetime.datetime(2024, 1, 1, 9, 0, 0),
-        ...     type=EnumTripType.PASSENGER_TRIP,
+        ...     trip_type=EnumTripType.PASSENGER_TRIP,
         ...     level_of_loading=0.8,
         ... )
         >>> trip_instance.save()
@@ -856,7 +915,7 @@ class Trip(models.Model):
     arrival_time = models.DateTimeField(blank=False)
 
     # Is the Trip empty, i.e., without passengers
-    type = models.CharField(
+    trip_type = models.CharField(
         max_length=9, choices=EnumTripType.choices, default=EnumTripType.PASSENGER_TRIP
     )
 
@@ -946,7 +1005,7 @@ class EventType(models.TextChoices):
     DRIVING = "DRIVING"
     CHARGING_OPPORTUNITY = "CHARGING_OPPORTUNITY"
     CHARGING_DEPOT = "CHARGING_DEPOT"
-    SERCVICE = "SERVICE"
+    SERVICE = "SERVICE"
     STANDBY_DEPARTURE = "STANDBY_DEPARTURE"
     PRECONDITIONING = "PRECONDITIONING"